requirements.txt
allure-pytest==2.8.29
allure-python-commons==2.8.29
attrs==19.3.0
behave==1.2.6
certifi==2019.11.28
chardet==3.0.4
enum34==1.1.10
essential-generators==0.9.2
idna==2.9
lxml==4.6.2
more-itertools==8.2.0
namedlist==1.8
packaging==20.3
parse==1.18.0
parse-type==0.5.2
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.6
pytest==6.2.2
requests==2.23.0
six==1.15.0
urllib3==1.25.8
wcwidth==0.1.9
